fix: satisfy session history lint boundaries
